carbonTrack/requirements.txt
asgiref==3.8.1
backports-zoneinfo==0.2.1
certifi==2025.8.3
charset-normalizer==3.4.3
django==4.2.24
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
mqtt==0.0.1
paho-mqtt==2.1.0
pillow==10.4.0
pyjwt==2.9.0
requests==2.32.4
sqlparse==0.5.3
typing-extensions==4.13.2
urllib3==2.2.3
